refactor: merge worktreeService.hooks.test.ts into worktreeService.test.ts
Issue: Test files were unnecessarily split for worktree hook functionality

Reason: Having separate test files for the same service creates maintenance
overhead and makes it harder to understand the full test coverage

Solution:
- Merged all hook execution tests into the main worktreeService.test.ts file
- Added necessary mocks for configurationManager and HookExecutor
- Added default mock for getWorktreeHooks to prevent existing test failures
- Removed the separate worktreeService.hooks.test.ts file

All tests continue to pass with the same coverage and functionality.
